requirements.txt
# Jarvis AI Dependencies - Updated 2026-03-23
# Python 3.12 Compatible - NO COMPILATION REQUIRED
fastapi==0.115.0
uvicorn[standard]==0.32.1
gunicorn==23.0.0
pydantic==2.10.6
python-multipart==0.0.18
edge-tts==6.1.15
gtts==2.5.4
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# Gmail API Integration
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.114.0

# Supabase Integration
supabase>=2.0.0
python-jose>=3.3.0
